pyproject.toml
[project]
name = "ztransfer"
version = "0.1.0"
description = "Local-first file transfer service"
readme = "PROJECT_PLAN.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "python-dotenv",
    "pydantic-settings",
    "python-multipart",
    "sqlmodel",
    "jinja2",
    "argon2-cffi",
    "itsdangerous",
    "email-validator",
    "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pytest",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_any_generics = true
